// src/app/features/nfse/pages/lista-notas-fiscais/models/duplicar-nota-state.ts
export interface DuplicarNotaState {
  serieRps?: string;
  tipoRps?: string;
  statusRps?: string;
  tributacaoRps?: string;
  discriminacao?: string;
  codigoMunicipio?: string;
  exigibilidadeISS?: string;
  municipioIncidencia?: string;
  tomadorCpfCnpj?: string;
  tomadorRazaoSocial?: string;
  tomadorEmail?: string;
  tomadorInscricaoMunicipal?: string;
  tomadorCep?: string;
  tomadorLogradouro?: string;
  tomadorNumero?: string;
  tomadorComplemento?: string;
  tomadorBairro?: string;
  tomadorCodigoMunicipio?: string;
  tomadorUf?: string;
  codigoServico?: string;
  valorServicos?: string;
  aliquotaServicos?: string;
  issRetido?: string;
  valorIss?: string;
  valorDeducoes?: string;
  valorPis?: string;
  valorCofins?: string;
  valorInss?: string;
  valorIr?: string;
  valorCsll?: string;
  outrasRetencoes?: string;
  descontoCondicionado?: string;
  descontoIncondicionado?: string;
  ibsIndDest?: string;
  ibsCstIbs?: string;
  ibsAliqEstadual?: string;
  ibsAliqMunicipal?: string;
  ibsCstCbs?: string;
  ibsAliqCbs?: string;
}